accept this result as bug fixed
We still have the number format type in ScFormulaCell. Removing that one
from ScFormulaCell is more work and needs much more inspection than
removing the formula index. Additionally we should inspect all callers
of ScFormulaCell::GetStandardFormat as most of them are useless when we
know that a cell has a fixed number format.

Change-Id: I3ad55a0c072da0fc4046d6c7245ed734a1dcaabc