KDE4: evaluate frameWidth in the Qt thread
QWidget::ensurePolished() uses events, which need to be processed in
the Qt thread. This happens when opening a document via Java UNO,
and no LO instance is already running. The Java UNO call is processed
in the main / Qt thread while loading happens in a second thread.

Actually getNativeControlRegion should not involve any drawing, but
KDE4 needs it to get the correct frame width.

Change-Id: I344d5089d958963c48a9a8a84bfa9fe8f092b75a