Implement SwPageDescs like o3tl::sorted_vector API
This moves the SwPageDescs class to pagedesc.hxx and extends
it from a vector class to a o3tl::sorted_vector compatible API.

This also drops multiple implementations for the "find SwPageDesc
by name algorithm".

This should just be a cleanup patch and is a preparation for the
change of SwPageDescs from vector to o3tl::sorted_vector.

Change-Id: I08561c139d1af9bad6ca68be27466fac33fab384